# lite_agent_web/lib/lite_agent_openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "web",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/model/add": {
      "post": {
        "summary": "新建模型",
        "deprecated": false,
        "description": "",
        "tags": [
          "ModelController"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelVOAddAction",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/model/{id}": {
      "delete": {
        "summary": "删除模型",
        "deprecated": false,
        "description": "",
        "tags": [
          "ModelController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "模型id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseData",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/model/update": {
      "put": {
        "summary": "修改模型",
        "deprecated": false,
        "description": "",
        "tags": [
          "ModelController"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelVOUpdateAction",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/model/list": {
      "get": {
        "summary": "模型列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "ModelController"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageNo",
            "in": "query",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataPageModelModelDTO",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": {
                        "pageNo": 0,
                        "pageSize": 0,
                        "total": 0,
                        "list": [
                          {
                            "id": "",
                            "name": "",
                            "baseUrl": "",
                            "apiKey": "",
                            "userId": "",
                            "workspaceId": "",
                            "shareFlag": false,
                            "createTime": "",
                            "updateTime": ""
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResponseData": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "id"
                },
                "name": {
                  "type": "string",
                  "description": "名称"
                },
                "userId": {
                  "type": "string",
                  "description": "创建者id"
                },
                "workspaceId": {
                  "type": "string",
                  "description": "工作空间id"
                },
                "icon": {
                  "type": "string",
                  "description": "图标"
                },
                "description": {
                  "type": "string",
                  "description": "描述"
                },
                "shareFlag": {
                  "type": "boolean",
                  "description": "分享标识",
                  "default": false
                },
                "dataSourceType": {
                  "type": "string",
                  "description": "数据源类型: INPUT, FILE, HTML"
                },
                "vectorCollectionName": {
                  "type": "string",
                  "description": "向量库集合名称"
                },
                "llmModelId": {
                  "type": "string",
                  "description": "关联的大模型id"
                },
                "embeddingModel": {
                  "type": "string",
                  "description": "embedding 模型"
                },
                "embeddingModelProvider": {
                  "type": "string",
                  "description": "embedding 模型提供商"
                },
                "retrievalTopK": {
                  "type": "integer",
                  "description": "检索 TopK"
                },
                "retrievalScoreThreshold": {
                  "type": "number",
                  "description": "检索分数阈值"
                },
                "apiUrl": {
                  "type": "string",
                  "description": "分享给外部的apiUrl"
                },
                "apiKey": {
                  "type": "string",
                  "description": "分享给外部的apikey"
                },
                "createTime": {
                  "type": "string",
                  "description": "创建时间",
                  "x-apifox-mock": "@datetime"
                },
                "updateTime": {
                  "type": "string",
                  "description": "更新时间",
                  "x-apifox-mock": "@datetime"
                }
              }
            },
            "description": "响应数据对象"
          }
        }
      },
...
      "ResponseDataPageModelModelDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "$ref": "#/components/schemas/PageModelModelDTO",
            "description": "响应数据对象"
          }
        }
      },
      "PageModelModelDTO": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "description": "当前第几页"
          },
          "pageSize": {
            "type": "integer",
            "description": "一页多少条记录"
          },
          "total": {
            "type": "integer",
            "description": "全部数据量",
            "format": "int64"
          },
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelDTO",
              "description": ""
            },
            "description": "数据"
          }
        }
      },
      "ModelDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": ""
          },
          "name": {
            "type": "string",
            "description": "模型名称"
          },
          "baseUrl": {
            "type": "string",
            "description": "模型访问url"
          },
          "apiKey": {
            "type": "string",
            "description": "模型访问key"
          },
          "userId": {
            "type": "string",
            "description": "创建者id"
          },
          "createUser": {
            "type": "string",
            "description": "创建人名称"
          },
          "maxTokens": {
            "type": "integer",
            "description": "限制最大token"
          },
          "type": {
            "type": "string",
            "description": "模型类型: text, embedding, image, video...",
            "default": "text"
          },
          "canEdit": {
            "type": "boolean",
            "description": "是否能编辑",
            "default": false
          },
          "canDelete": {
            "type": "boolean",
            "description": "是否能删除",
            "default": false
          },
          "canRead": {
            "type": "boolean",
            "description": "是否能读",
            "default": false
          }
        }
      },
...
      "ModelVOUpdateAction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": ""
          },
          "name": {
            "type": "string",
            "description": "模型名字"
          },
          "baseUrl": {
            "type": "string",
            "description": "请求url"
          },
          "apiKey": {
            "type": "string",
            "description": "接口key"
          },
          "maxTokens": {
            "type": "integer",
            "description": "限制最大token",
            "minimum": 1,
            "x-apifox-mock": "@integer(1)"
          },
          "type": {
            "type": "string",
            "description": "模型类型: text, embedding, image, video..."
          }
        },
        "required": [
          "id",
          "name",
          "baseUrl",
          "apiKey",
          "type"
        ]
      },
...
      "ModelVOAddAction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": ""
          },
          "name": {
            "type": "string",
            "description": "模型名字"
          },
          "baseUrl": {
            "type": "string",
            "description": "请求url"
          },
          "apiKey": {
            "type": "string",
            "description": "接口key"
          },
          "maxTokens": {
            "type": "integer",
            "description": "限制最大token",
            "minimum": 1,
            "x-apifox-mock": "@integer(1)"
          },
          "type": {
            "type": "string",
            "description": "模型类型: text, embedding, image, video..."
          }
        },
        "required": [
          "name",
          "baseUrl",
          "apiKey",
          "type"
        ]
      },
...
      "ResponseDataString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "type": "string",
            "description": "响应数据对象"
          }
        }
      }
    },
    "securitySchemes": {}
  },
  "servers": []
}